% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGroups.r
\name{all_combinations}
\alias{all_combinations}
\title{Compute the number of all possible set assignments}
\usage{
all_combinations(x)
}
\arguments{
\item{x}{A vector of N set sizes}
}
\value{
The number of all possible assignments to N sets
}
\description{
Compute the number of all possible set assignments
}
