% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minDiff-package.R
\docType{package}
\name{minDiff-package}
\alias{minDiff-package}
\title{The minDiff Package}
\description{
Minimize differences between groups.
}
\details{
\tabular{ll}{
Package: \tab minDiff\cr
Type: \tab Package\cr
Version: \tab 0.01-2\cr
Date: \tab 2018-02-02\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab https://github.com/m-Py/minDiff\cr
}

Assign any sort of elements to different groups and minimize group differences
}
\author{
Martin Papenberg
}
\keyword{package}
